backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
google-cloud-alloydb-connector
llama-index
llama-parse
google-generativeai
google-cloud-monitoring>=2.18.0
structlog
hypothesis
pytest
pytest-asyncio
httpx
aiosqlite
slowapi
tenacity
